Fail MP3 probe if AC3 or EC3 detected Related to #4958
